its-camera-ai/pyproject.toml
[project]
name = "its-camera-ai"
version = "0.1.0"
description = "AI Camera Traffic Monitoring System - Real-time traffic analytics and vehicle tracking"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "ITS Camera AI Team", email = "team@its-camera-ai.com"},
]
keywords = ["ai", "computer-vision", "traffic-monitoring", "pytorch", "fastapi"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Core production dependencies
dependencies = [
    # Web Framework & API
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35.0",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.10.1",
    # Async & Concurrency
    "asyncio-mqtt>=0.16.2",
    "aioredis>=2.0.1",
    "aiokafka>=0.12.0",
    "httpx>=0.28.1",
    # Computer Vision & ML
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "ultralytics>=8.3.177",
    "opencv-python-headless>=4.8.1",
    "numpy<2.2.5",
    "pillow>=11.3.0",
    # Data Processing
    "pandas>=2.3.1",
    "pyarrow>=21.0.0",
    "polars>=1.32.2",
    # Database & Storage
    "sqlalchemy[asyncio]>=2.0.43",
    "asyncpg>=0.30.0",
    "redis>=6.4.0",
    "minio>=7.2.16",
    # Time Series & Analytics
    "influxdb-client[async]>=1.49.0",
    "prometheus-client>=0.22.1",
    # Security & Authentication
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "cryptography>=42.0.0",
    "PyJWT>=2.8.0",
    "bcrypt>=4.1.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
    # Configuration & Environment
    "python-dotenv>=1.1.1",
    "structlog>=25.4.0",
    "rich>=14.1.0",
    # Monitoring & Observability
    "opentelemetry-api>=1.36.0",
    "opentelemetry-sdk>=1.36.0",
    "opentelemetry-instrumentation-fastapi>=0.57b0",
    "sentry-sdk[fastapi]>=2.34.1",
]

[dependency-groups]
dev = [
    # Type Checking & Linting
    "mypy>=1.17.1",
    "ruff>=0.12.8",
    "black>=25.1.0",
    "isort>=6.0.1",
    
    # Testing
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "pytest-benchmark>=5.1.0",
    "hypothesis>=6.137.2",
    
    # Development Tools
    "pre-commit>=4.3.0",
    "bandit[toml]>=1.8.6",
    "safety>=3.6.0",
    "pip-audit>=2.9.0",
    
    # Documentation
    "sphinx>=8.2.3",
    "sphinx-rtd-theme>=3.0.2",
    "myst-parser>=4.0.1",
    
    # Type Stubs
    "types-passlib>=1.7.7.20250602",
    "types-redis>=4.6.0.20241004",
    "types-pillow>=10.2.0.20240822",
    "types-setuptools>=80.9.0.20250809",
]

ml = [
    # Extended ML Dependencies
    "scikit-learn>=1.7.1",
    "xgboost>=3.0.4",
    "optuna>=4.4.0",
    "mlflow>=3.2.0",
    "tensorboard>=2.20.0",
    "onnx>=1.18.0",
    "onnxruntime-gpu>=1.22.0",
]

gpu = [
    # GPU-specific dependencies
    "triton>=3.4.0",
    "cupy-cuda12x>=13.5.1",
]

linux = [
    # Linux-specific dependencies
    # "torch-tensorrt>=1.4.0",
]

edge = [
    # Edge deployment dependencies
    "optimum[onnxruntime]>=1.27.0",
    "openvino>=2025.2.0",
    "pycoral>=0.2.0",
]

[project.scripts]
its-camera-ai = "its_camera_ai.cli:main"

[project.urls]
Homepage = "https://github.com/your-org/its-camera-ai"
Documentation = "https://its-camera-ai.readthedocs.io"
Repository = "https://github.com/your-org/its-camera-ai.git"
Issues = "https://github.com/your-org/its-camera-ai/issues"

# Build System Configuration
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

# Tool Configuration
[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["its_camera_ai*"]

# Black Configuration
[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | \.venv
    | build
    | dist
    | migrations
)/
'''

# isort Configuration
[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

# Ruff Configuration
[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "S",   # bandit security checks
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "S101",  # assert used (OK in tests)
]
exclude = [
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".venv",
    "build",
    "dist",
    "migrations",
]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101", "ARG002", "ARG001"]

# MyPy Configuration
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
extra_checks = true

[[tool.mypy.overrides]]
module = [
    "cv2.*",
    "ultralytics.*",
    "kafka.*",
    "redis.*",
    "minio.*",
    "influxdb_client.*",
]
ignore_missing_imports = true

# Pytest Configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=its_camera_ai",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--cov-fail-under=90",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "gpu: marks tests that require GPU",
    "ml: marks tests for ML models",
]

# Coverage Configuration
[tool.coverage.run]
source = ["src/its_camera_ai"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

# Bandit Security Configuration
[tool.bandit]
exclude_dirs = ["tests", "migrations", ".venv"]
skips = ["B101", "B601"]  # Skip assert_used and shell_injection in tests

[tool.bandit.assert_used]
skips = ["*/test_*.py", "*/tests.py"]
